Necessary edits for Swagger OAuth2 support
* additional arity is added to fnhouse.swagger/collect-routes. New
parameter is a function from handler's annotation to swagger's metadata.
It complements fnhouse's extra-info-fn, allowing to control Swagger
spec from a code that uses fnhouse-swagger.
* swagger-parameters map is now expected in resources. It will be
passed to ring-swagger as-is.

NOTE: this is a breaking change, because swagger-parameters map should
be provided in resources. fnk's optional notation ({foo "bar"})
doesn't seem to work in parameters map, possibly due to fnhouse or fnk
bug.